requirements.txt
google-generativeai
chromadb
tqdm
pandas
python-dotenv
sentence-transformers
streamlit